packages/sayou-document/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# -----------------
# 1. 패키지 기본 정보
# -----------------
[project]
name = "sayou-document"
version = "0.0.1"
authors = [
    { name = "Sayouzone", email = "contact@sayouzone.com" },
]
description = "Document components for the Sayou Data Platform"
readme = "README.md"
license = { file = "../../LICENSE" }
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]
dependencies = [
    "sayou-core ~= 0.0.3",
    "openpyxl==3.1.5",
    "PyMuPDF==1.25.5",
    "python-docx==1.1.2",
    "python-pptx==1.0.2",
    "pydantic ~= 2.12.4"
]

# -----------------
# 2. 프로젝트 링크 (PyPI 사이드바)
# -----------------
[project.urls]
"Homepage" = "https://www.sayouzone.com/"
"Documentation" = "https://sayouzone.github.io/sayou-fabric/"
"Repository" = "https://github.com/sayouzone/sayou-fabric"

# -----------------
# 3. 소스 코드 위치
# -----------------
[tool.hatch.build.targets.wheel]
packages = ["src/sayou"]